+++ bathysphere.yaml
@@ -328,6 +328,21 @@ paths:
         '400':
           $ref: '#/components/responses/BadRequest'
 
+  /storage:
+    get:
+      tags: [Storage]
+      summary: Retrieve arbitrary S3 data
+      security:
+        - BearerAuth: []
+      description: Retrieve arbitrary S3 data
+      parameters: []
+      responses:
+        '200':
+          description: Success
+        '404':
+          $ref: '#/components/responses/NotFound'
+
+
 components:
 
   securitySchemes:
